+++ lib/sec_id.rb
@@ -119,7 +119,6 @@ module SecId
 end
 
 require 'sec_id/validation_result'
-require 'sec_id/concerns/normalizable'
 require 'sec_id/concerns/checkable'
 require 'sec_id/base'
 require 'sec_id/detector'